packages/backend/requirements.txt
fastapi[standard]
psycopg[binary]
sqlmodel
passlib
bcrypt
pyjwt
minio
python-multipart
Pillow
pydantic-settings
uvicorn